+++ packages/collector/test/tracing/cloud/aws/sqs/sqsUtil.js
@@ -1,0 +1,14 @@
+/*
+ * (c) Copyright IBM Corp. 2021
+ * (c) Copyright Instana Inc. and contributors 2021
+ */
+
+'use strict';
+
+const AWS = require('aws-sdk');
+
+AWS.config.update({ region: 'us-east-2' });
+
+const sqs = new AWS.SQS();
+
+exports.sqs = sqs;