+++ packages/core/src/tracing/index.js
@@ -25,6 +25,7 @@ let processIdentityProvider = null;
 // Note: Also update initializedTooLateHeuristic.js and the accompanying test when adding instrumentations.
 const instrumentations = [
   './instrumentation/cloud/aws/sdk',
+  './instrumentation/cloud/aws/sqs',
   './instrumentation/cloud/gcp/pubsub',
   './instrumentation/cloud/gcp/storage',
   './instrumentation/control_flow/bluebird',